TTS of playht and openai
Created a script for both AI. The API key is in a separate file and ignored by git, due to security issues. FYI the text at the moment is set and is not dynamic. To Do: merge with Anton's script (openai) to generate the speech dynamically, based on openai's text.